+++ src/pages/StockNurYadi.tsx
@@ -1028,7 +1028,7 @@ function StockNurYadiInner() {
             <div className="mb-6">
               <div className="flex items-end justify-between">
                 <div>
-                <h1 className="text-[11px] font-normal tracking-[0.2em] uppercase text-dim mb-1">{new Date().toLocaleDateString("en-GB", { weekday: "short", day: "numeric", month: "short" })}</h1>
+                <h1 className="text-[11px] font-normal tracking-[0.2em] uppercase text-dim mb-1">{new Date().toLocaleDateString("en-GB", { weekday: "short", day: "numeric", month: "long" })}</h1>
                   <p className="text-[28px] font-light tracking-tight">Nur Yadi</p>
                 </div>
                 <span
